+++ sdk/ai/ai-projects/CHANGELOG.md
@@ -1,11 +1,13 @@
 # Release History
 
-## 1.0.0 (2025-07-08)
+## 1.0.0 (2025-07-09)
 
 ### Breaking changes
 
 - remove `project.redTeams`
 - remove `project.evaluations`
+- remove `project.inference.chatCompletions`, `project.inference.embeddings` and `project.inference.imageEmbeddings`. For guidance on obtaining an authenticated `azure-ai-inference` client for your AI Foundry Project,
+refer to the updated samples in the `samples\inference` directory.
 
 ## 1.0.0-beta.10 (2025-07-02)